폴더구조 변경 필요없는 것들 1차 삭제
