--- src/main/java/com/ttallang/user/TtallangApplication.java
@@ -1,13 +1,0 @@
-package com.ttallang.user;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TtallangApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TtallangApplication.class, args);
-    }
-
-}
