+++ src/main/java/com/ttallang/user/TtallangUserApplication.java
@@ -1,0 +1,13 @@
+package com.ttallang.user;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TtallangUserApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TtallangUserApplication.class, args);
+    }
+
+}
