--- src/main/java/com/ttallang/user/userAuth/security/config/handler/OAuth2Handler.java
@@ -1,23 +1,0 @@
-package com.ttallang.user.userAuth.security.config.handler;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-
-import java.io.IOException;
-
-public class OAuth2Handler implements AuthenticationSuccessHandler {
-
-    @Override
-    public void onAuthenticationSuccess(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            Authentication authentication
-    ) throws IOException {
-        // 인증 성공 처리.
-        OAuth2User oAuth2User = (OAuth2User) authentication.getPrincipal();
-        response.sendRedirect("/");  // 원하는 페이지로 변경
-    }
-}
