--- src/main/java/com/ttallang/user/userAuth/security/repository/RolesRepository.java
@@ -1,13 +1,0 @@
-package com.ttallang.user.userAuth.security.repository;
-
-import com.ttallang.user.commonModel.Roles;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-//@Repository 상속 없이 ioc 가능 JpaRepository 상속해서 자동 bean 등록
-public interface RolesRepository extends JpaRepository<Roles, Integer> {
-
-//    @Query("")
-//    User findByUserId
-    Roles findByUserName(String userName);
-    void deleteByUserName(String userName);
-}
