--- src/main/java/com/ttallang/user/userAuth/security/service/SecurityService.java
@@ -1,8 +1,0 @@
-package com.ttallang.user.userAuth.security.service;
-
-import java.util.Map;
-
-public interface SecurityService {
-    void signupCustomer(Map<String, String> userData);
-    boolean isExistingCustomer(String userName);
-}
