--- src/main/webapp/WEB-INF/views/intro.jsp
@@ -1,21 +1,0 @@
-<%--
-  Created by IntelliJ IDEA.
-  User: KOSA
-  Date: 2024-10-28
-  Time: 오후 3:58
-  To change this template use File | Settings | File Templates.
---%>
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<html>
-<head>
-	<title>Intro</title>
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
-</head>
-<body>
-인트로 페이지...
-<a href="/loginForm" type="button" class="btn btn-primary">로그인</a>
-<br>
-<a href="/signupForm" type="button" class="btn btn-primary">회원가입</a>
-</body>
-</html>
