--- src/main/webapp/WEB-INF/views/userAuth/main/paycoResult.jsp
@@ -1,16 +1,0 @@
-<%--
-  Created by IntelliJ IDEA.
-  User: mhd32
-  Date: 24. 11. 3.
-  Time: 오후 7:56
-  To change this template use File | Settings | File Templates.
---%>
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<html>
-<head>
-  <title>Title</title>
-</head>
-<body>
-  ${userInfo}
-</body>
-</html>
