+++ src/main/webapp/WEB-INF/views/userAuth/test/adminPage.jsp
@@ -1,0 +1,18 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: KOSA
+  Date: 2024-11-04
+  Time: 오후 2:13
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+<head>
+	<title>Title</title>
+</head>
+<body>
+어드민 페이지...
+유저 아이디: ${userId}
+유저 네임: ${username}
+</body>
+</html>
