--- src/main/webapp/WEB-INF/views/userAuth/test/loginSuccess.jsp
@@ -1,17 +1,0 @@
-<%--
-  Created by IntelliJ IDEA.
-  User: KOSA
-  Date: 2024-10-30
-  Time: 오후 3:10
-  To change this template use File | Settings | File Templates.
---%>
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<html>
-<head>
-	<title>Title</title>
-</head>
-<body>
-로그인 성공!!!
-customerId: ${customerId}
-</body>
-</html>
