--- src/main/webapp/WEB-INF/views/userAuth/test/myPage.jsp
@@ -1,17 +1,0 @@
-<%--
-  Created by IntelliJ IDEA.
-  User: KOSA
-  Date: 2024-10-31
-  Time: 오전 10:39
-  To change this template use File | Settings | File Templates.
---%>
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<html>
-<head>
-	<title>Title</title>
-</head>
-<body>
-프로필 페이지?
-customerId: ${customerId}
-</body>
-</html>
